Changed file names to index
