+++ src/Header/index.js
@@ -1,3 +1,7 @@
-import Header from './Header'
+import React from 'react';
+
+import './index.css';
+
+const Header = () => <header className="header">Header</header>;
 
 export default Header;
